fix(app): address persistence integration edge cases
