+++ app/lib/l10n/app_de.arb
@@ -516,6 +516,8 @@
   "recovery_failed": "Wiederherstellung fehlgeschlagen. Bitte überprüfe deine Wiederherstellungsphrase und versuche es erneut.",
   "recovery_user_not_found": "Kein Konto mit dieser Wiederherstellungsphrase gefunden.",
   "recovery_network_error": "Netzwerkfehler. Bitte überprüfe deine Verbindung und versuche es erneut.",
+  "recovery_cleanup_failed": "Die Wiederherstellung war erfolgreich, aber dieses Gerät konnte die lokale Bereinigung nicht abschließen. Bitte versuche es erneut.",
+  "recovery_local_persistence_failed": "Die Wiederherstellung war erfolgreich, aber dieses Gerät konnte deine Kontodaten nicht lokal speichern. Bitte versuche es erneut.",
 
   "restore_account_description": "Stelle dein Konto auf diesem Gerät mit der Wiederherstellungsphrase wieder her, die du beim Beitritt zur Studie gespeichert hast.",
   "restore_account_help_title": "Mit Wiederherstellungsphrase wiederherstellen",
