+++ app/lib/screens/app_onboarding/restore_account_screen.dart
@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:provider/provider.dart';
 import 'package:studyu_app/app_router.dart';
 import 'package:studyu_app/l10n/app_localizations.dart';
+import 'package:studyu_app/services/authenticated_language_sync_service.dart';
 import 'package:studyu_app/services/restore_account_service.dart';
 import 'package:studyu_app/widgets/onboarding_page.dart';
 import 'package:studyu_core/core.dart';
-import 'package:studyu_flutter_common/studyu_flutter_common.dart';
 
 class RestoreAccountScreen extends StatefulWidget {
   const RestoreAccountScreen({super.key});
@@ -58,6 +57,10 @@ class _RestoreAccountScreenState extends State<RestoreAccountScreen> {
         return localizations.recovery_user_not_found;
       case 'recovery_network_error':
         return localizations.recovery_network_error;
+      case 'recovery_cleanup_failed':
+        return localizations.recovery_cleanup_failed;
+      case 'recovery_local_persistence_failed':
+        return localizations.recovery_local_persistence_failed;
       case 'recovery_failed':
         return localizations.recovery_failed;
       default:
@@ -114,7 +117,7 @@ class _RestoreAccountScreenState extends State<RestoreAccountScreen> {
       }
 
       if (!mounted) return;
-      await context.read<AppLanguage>().syncWithAuthenticatedUser();
+      await syncLanguageForAuthenticatedUser(context);
       if (!mounted) return;
 
       if (result.subjectId != null) {
